requirements.txt
torch==1.13.1
torchaudio==0.13.1
pyannote.audio==2.1.1
pydub==0.25.1
SpeechRecognition==3.10.0
transformers==4.38.0
numpy==1.26.4
soundfile>=0.10.2,<0.11
huggingface_hub==0.23.0